pyproject.toml
[tool.poetry]
name = "uchannel"
version = "1.0.0"
description = "Create an infinite number of structural U-shaped CAD sheet metals using four distinct parametric models"
authors = ["Tobias Lehrer <tobias.lehrer@tum.de>", "Philipp Stocker <philipp.stocker@oth-regensburg.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gmsh = "^4.13.1"
numpy = "^2.1.3"
pyDOE3 = "^1.0.4"
pillow = "^11.0.0"
joblib = "^1.4.2"
PyYAML = "^6.0.2"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
